tweaking
Trying to run the scripts wihtout the info_20 table (due to the EOF error), as only that file causes the error. Going through the rest of the analysis scripts to make sure the rest makes sense. Some things don't make sense atm... mainly the subsetting for the analysis scripts.... think to do with model_ids